* commit 'e859a34171f2a36877d95197d118d962078f8aa0':
  Rewrite PictureSet with TURBO!

+++ Source/WebCore/platform/graphics/android/rendering/TilesManager.h
@@ -87,8 +87,8 @@ public:
     int currentLayerTextureCount();
     void setCurrentTextureCount(int newTextureCount);
     void setCurrentLayerTextureCount(int newTextureCount);
-    static float tileWidth();
-    static float tileHeight();
+    static int tileWidth();
+    static int tileHeight();
 
     void allocateTextures();
